version 0.0.7 of NeuroTensor

+++ src/permute/permute.cpp
@@ -12,22 +12,22 @@ namespace nt{
 namespace permute{
 
 
-Permuter::Permuter(void** s, const std::vector<uint32_t>& st, const std::vector<uint32_t>& sh)
+Permuter::Permuter(void** s, const std::vector<int64_t>& st, const std::vector<int64_t>& sh)
 	:strides_old(s),
 	strides(st),
 	shape(sh),
 	shape_accumulate(sh.size())
 {
-	for(uint32_t i = 0; i < shape_accumulate.size(); ++i){
-		shape_accumulate[i] = std::accumulate(shape.cbegin()+i+1, shape.cend(), 1, std::multiplies<uint32_t>());
+	for(int64_t i = 0; i < shape_accumulate.size(); ++i){
+		shape_accumulate[i] = std::accumulate(shape.cbegin()+i+1, shape.cend(), 1, std::multiplies<int64_t>());
 	}
 }
 
-uint32_t Permuter::get_index(uint32_t idx) const {
-	uint32_t n_index = 0;
-	for(uint32_t i = 0; i < shape.size(); ++i){
+int64_t Permuter::get_index(int64_t idx) const {
+	int64_t n_index = 0;
+	for(int64_t i = 0; i < shape.size(); ++i){
 		if(idx >= shape_accumulate[i]){
-			uint32_t idx_n = idx / shape_accumulate[i];
+			int64_t idx_n = idx / shape_accumulate[i];
 			idx = idx % shape_accumulate[i];
 			n_index += idx_n * strides[i];
 		}
@@ -35,11 +35,11 @@ uint32_t Permuter::get_index(uint32_t idx) const {
 	return n_index;
 }
 
-void* Permuter::get_ptr(uint32_t idx){
+void* Permuter::get_ptr(int64_t idx){
 	void** o_copy = strides_old;
-	for(uint32_t i = 0; i < shape.size() && idx > 0; ++i){
+	for(int64_t i = 0; i < shape.size() && idx > 0; ++i){
 		if(idx >= shape_accumulate[i]){
-			uint32_t idx_n = idx / shape_accumulate[i];
+			int64_t idx_n = idx / shape_accumulate[i];
 			idx = idx % shape_accumulate[i];
 			o_copy += idx_n * strides[i];
 		}
@@ -48,9 +48,9 @@ void* Permuter::get_ptr(uint32_t idx){
 }
 
 
-void Permute(void** ar, void** n_str, uint32_t size, const std::vector<uint32_t>& new_shape, const std::vector<uint32_t>& new_strides){
+void Permute(void** ar, void** n_str, int64_t size, const std::vector<int64_t>& new_shape, const std::vector<int64_t>& new_strides){
 	Permuter perm(ar, new_strides, new_shape);
-	for(uint32_t i = 0; i < size; ++i, ++n_str)
+	for(int64_t i = 0; i < size; ++i, ++n_str)
 		*(n_str) = perm.get_ptr(i);
 }
